requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
aiortc==1.9.0
aiofiles==23.2.1
pyyaml==6.0.1
numpy==1.26.4
av==12.0.0
pyee==11.1.0
psutil==5.9.8
python-multipart==0.0.9
